+++ src/felupe/mesh/_tools.py
@@ -303,8 +303,8 @@ def revolve(points, cells, cell_type, n=11, phi=180, axis=0):
 
 
 @mesh_or_data
-def sweep(points, cells, cell_type, decimals=None):
-    """Merge duplicated points and update cells of a Mesh.
+def merge_duplicate_points(points, cells, cell_type, decimals=None):
+    """Merge duplicate points and update cells of a Mesh.
 
     Parameters
     ----------
@@ -332,7 +332,7 @@ def sweep(points, cells, cell_type, decimals=None):
         This function re-sorts points.
 
     ..  note::
-        This function does not merge duplicated cells.
+        This function does not merge duplicate cells.
 
     Examples
     --------
@@ -418,6 +418,113 @@ def sweep(points, cells, cell_type, decimals=None):
 
     return points_new, cells_new, cell_type
 
+
+@mesh_or_data
+def merge_duplicate_cells(points, cells, cell_type):
+    """Merge duplicate cells of a Mesh.
+    
+    Parameters
+    ----------
+    points : list or ndarray
+        Original point coordinates.
+    cells : list or ndarray
+        Original point-connectivity of cells.
+    cell_type : str
+        A string in VTK-convention that specifies the cell type.
+    
+    Returns
+    -------
+    points : ndarray
+        Point coordinates.
+    cells : list or ndarray
+        Cells with merged duplicate cells.
+    cell_type : str or None
+        A string in VTK-convention that specifies the cell type.
+    
+    Notes
+    -----
+    ..  warning::
+        This function re-sorts cells.
+
+    ..  note::
+        This function does not merge duplicate points.
+    
+    Examples
+    --------
+    Two quad meshes to be merged overlap some cells. Merge these duplicated
+    points and update the cells.
+
+    >>> import felupe as fem
+
+    >>> rect1 = fem.Rectangle(n=11)
+    >>> rect2 = fem.Rectangle(a=(0.9, 0), b=(1.9, 1), n=11)
+    >>> rect2
+    <felupe Mesh object>
+      Number of points: 121
+      Number of cells:
+        quad: 100
+
+    Each mesh contains 121 points and 100 cells. These two meshes are now stored in a
+    :class:`~felupe.MeshContainer`.
+
+    >>> container = fem.MeshContainer([rect1, rect2])
+    >>> container
+    <felupe mesh container object>
+      Number of points: 242
+      Number of cells:
+        quad: 100
+        quad: 100
+
+    The meshes of the mesh container are :func:`stacked <felupe.mesh.stack>`.
+
+    >>> stack = fem.mesh.stack(container.meshes)
+    >>> stack
+    <felupe Mesh object>
+      Number of points: 242
+      Number of cells:
+        quad: 200
+
+    After merging the duplicated points and cells, the number of points is reduced but
+    the number of cells is unchanged.
+
+    >>> mesh = fem.mesh.merge_duplicate_points(stack)
+    >>> mesh
+    <felupe Mesh object>
+      Number of points: 220
+      Number of cells:
+        quad: 200
+    
+    >>> ax = mesh.imshow(opacity=0.6)
+    
+    ..  image:: images/mesh_sweep.png
+        :width: 400px
+    
+    ..  note::
+        The :class:`~felupe.MeshContainer` may be directly created with ``merge=True``.
+        This enforces :func:`~felupe.mesh.merge_duplicate_points` for the shared points
+        array of the container.
+    
+    The duplicate cells are merged in a second step.
+    
+    >>> merged = fem.mesh.merge_duplicate_cells(mesh)
+    >>> merged
+    <felupe Mesh object>
+      Number of points: 220
+      Number of cells:
+        quad: 190
+    
+    ..  image:: images/mesh_merged.png
+        :width: 400px
+
+    See Also
+    --------
+    Mesh.merge_duplicate_points : Merge duplicate points of a Mesh.
+    Mesh.merge_duplicate_cells : Merge duplicate cells of a Mesh.
+    MeshContainer : A container which operates on a list of meshes with identical
+        dimensions.
+    """
+    
+    return points, np.unique(cells, axis=0), cell_type
 
 @mesh_or_data
 def translate(points, cells, cell_type, move, axis):